move parsing to own component, add hack for sidebar (works, but not pretty)
